refactor: type 해당 파일에서 선언
